requirements.txt
django
djangorestframework
django-storages
django-environ
boto3
pillow
psycopg2
